+++ CHANGELOG.md
@@ -4,6 +4,10 @@ documented here.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.9]
+  * Add the `ComplexField` trait.
+  * Rename the `Real` trait to `RealField` for more clarity.
+
 ## [0.5.2]
   * Add feature-gated support for the `decimal` crate.
   * Add attributes to automatically generate quickcheck checks.